Move physics math to physics/utils.py
