+++ pooltool/physics/__init__.py
@@ -35,6 +35,15 @@ from pooltool.physics.resolve.transition import (
     BallTransitionModel,
     ball_transition_models,
 )
+from pooltool.physics.utils import (
+    get_ball_energy,
+    get_roll_time,
+    get_slide_time,
+    get_spin_time,
+    get_u_vec,
+    rel_velocity,
+    surface_velocity,
+)
 
 __all__ = [
     "PhysicsEngine",
@@ -48,6 +57,13 @@ __all__ = [
     "BallPocketModel",
     "StickBallModel",
     "BallTransitionModel",
+    "rel_velocity",
+    "surface_velocity",
+    "get_u_vec",
+    "get_slide_time",
+    "get_roll_time",
+    "get_spin_time",
+    "get_ball_energy",
     "ball_ball_models",
     "BallBallFrictionModel",
     "ball_ball_friction_models",
