+++ pooltool/ptmath/__init__.py
@@ -10,23 +10,15 @@ from pooltool.ptmath.utils import (
     cross,
     decompose_normal_tangent,
     find_intersection_2D,
-    get_ball_energy,
-    get_roll_time,
-    get_slide_time,
-    get_spin_time,
-    get_u_vec,
     is_overlapping,
     norm2d,
     norm3d,
     point_on_line_closest_to_point,
     quaternion_from_vector_to_vector,
-    rel_velocity,
     rotation_from_vector_to_vector,
     solve_transcendental,
     squared_norm2d,
     squared_norm3d,
-    surface_velocity,
-    tangent_surface_velocity,
     unit_vector,
     unit_vector_slow,
     wiggle,
@@ -51,15 +43,7 @@ __all__ = [
     "unit_vector",
     "unit_vector_slow",
     "wiggle",
-    "rel_velocity",
     "rotation_from_vector_to_vector",
     "quaternion_from_vector_to_vector",
-    "surface_velocity",
-    "tangent_surface_velocity",
-    "get_u_vec",
-    "get_slide_time",
-    "get_roll_time",
-    "get_spin_time",
-    "get_ball_energy",
     "is_overlapping",
 ]
